Cross-compile with Dotty
Tests that do not compile or failed with Dotty were moved to
scala-2.x directories and should be investigated at some point.
