+++ .travis.yml
@@ -7,6 +7,7 @@ jdk:
 scala:
   - 2.12.8
   - 2.13.0-RC2
+  - 0.14.0-RC1
 
 env:
   global:
@@ -27,6 +28,10 @@ matrix:
     - jdk: openjdk11
       env: SCALAJS_VERSION=0.6.27
     - jdk: openjdk11
+      env: SCALAJS_VERSION=1.0.0-M7
+    - scala: 0.14.0-RC1
+      env: SCALAJS_VERSION=0.6.27
+    - scala: 0.14.0-RC1
       env: SCALAJS_VERSION=1.0.0-M7
 
 script:
