+++ jvm/src/test/scala-2.x/scala/xml/CompilerErrors.scala
