+++ shared/src/test/scala/scala/xml/ShouldCompile.scala
@@ -6,12 +6,6 @@ package scala.xml
 import scala.collection._
 import scala.collection.mutable.ArrayBuffer
 
-// t1626
-object o {
-  val n = <a xmlns=""/>
-  n.namespace == null
-}
-
 // t1761
 class Foo {
   val elements: Seq[Node] = Nil
